requirements.txt
python-dotenv
openai
sentence-transformers
faiss-cpu
streamlit
pypdf
tqdm
numpy
tqdm
rank-bm25
